requirements.txt
anthropic>=0.39.0
discord.py>=2.3.2
python-telegram-bot>=20.8
